+++ tests/test_taskgroups.py
@@ -539,7 +539,13 @@ async def test_cancel_scope_cleared() -> None:
 async def test_fail_after(delay: float) -> None:
     with pytest.raises(TimeoutError):
         with fail_after(delay) as scope:
-            await sleep(1)
+            try:
+                await sleep(1)
+            except get_cancelled_exc_class() as exc:
+                assert "deadline" in str(exc)
+                raise
+            else:
+                pytest.fail("sleep() should have raised a cancellation exception")
 
     assert scope.cancel_called
     assert scope.cancelled_caught
@@ -1798,3 +1804,15 @@ async def test_exception_groups_suppresses_exc_context() -> None:
             raise Exception("Error")
 
     assert exc_info.value.__suppress_context__
+
+
+async def test_cancel_reason() -> None:
+    with CancelScope() as scope:
+        scope.cancel("test reason")
+        with pytest.raises(get_cancelled_exc_class()) as exc_info:
+            await checkpoint()
+
+    task = get_current_task()
+    assert task and task.name
+    exc_info.match("test reason")
+    exc_info.match(task.name)